Code refactoring and clean up.
Bad or empty input dates are created as a new Date() rather than new Date(0,0,0,0,0) (or a variation thereof).  This means things are far less likely to break as you aren't going to try and draw a chart in hour format starting in 1899!

Optimised repeated code related to creation of captions.

Stopped dependency lines between tasks within a combined group from doing a loop when the start and end point are close together.

Removed some unnecessary variable assignment during table creation using calls to newNode.

Looking at the code with fresh eyes during the most recent updates I finally realised that when I refactored the code to use createElement() and appendChild() rather than innerHTML I removed the need to use JSGantt.findObj() to locate nodes in the document.  This lead to the following changes:
As all the nodes exist straight after the call to createElement all listeners are registered as soon as possible rather than as a group at the end.  As a result...
All listener creation methods were updated to use objects rather than strings/JSGantt.findObj().
Updated how/where dependency lines are created in the document.
Simplified the display of the tool tip.